# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Rokwire Groups Building Block API Documentation.",
        "title": "Rokwire Groups Building Block API",
        "contact": {},
        "version": "1.29.0"
    },
    "host": "localhost",
    "basePath": "/gr",
    "paths": {
...
        "/api/admin/group/events/v3": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Create a calendar event and link it to multiple group ids",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminCreateCalendarEventMultiGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventMultiGroupData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventMultiGroupData"
                        }
                    }
                }
            }
        },
...
        "/api/admin/group/{group-id}/events/v3": {
            "put": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Updates a calendar event and for a single group id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminUpdateCalendarEventSingleGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.updateCalendarEventSingleGroupData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.updateCalendarEventSingleGroupData"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Create a calendar event and link it to a single group id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminCreateCalendarEventSingleGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventSingleGroupData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventSingleGroupData"
                        }
                    }
                }
            }
        },
        "/api/admin/group/{group-id}/events/v3/load": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Gets the group calendar events",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminGetGroupCalendarEventsV3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/GroupEventFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/admin/user/event/{event-id}/groups": {
            "get": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Updates the group mappings for an event with id",
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminUpdateGroupMappingsEventID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "event-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/getPutAdminGroupIDsForEventIDRequestAndResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/getPutAdminGroupIDsForEventIDRequestAndResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/user/event/{event_id}/groups": {
            "get": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Get all group IDs where the current user is an admin",
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminGetAdminGroupIDsForEventID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "event-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/getPutAdminGroupIDsForEventIDRequestAndResponse"
                        }
                    }
                }
            }
        },
...
        "/api/group/events/v3": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Create a calendar event and link it to multiple group ids",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "CreateCalendarEventMultiGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventMultiGroupData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventMultiGroupData"
                        }
                    }
                }
            }
        },
...
        "/api/group/{group-id}/events/v3": {
            "put": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Updates a calendar event and for a single group id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "UpdateCalendarEventSingleGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.updateCalendarEventSingleGroupData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.updateCalendarEventSingleGroupData"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Create a calendar event and link it to a single group id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "CreateCalendarEventSingleGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventSingleGroupData"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.createCalendarEventSingleGroupData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "getPutAdminGroupIDsForEventIDRequestAndResponse": {
            "type": "object",
            "properties": {
                "group_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "rest.createCalendarEventMultiGroupData": {
            "type": "object",
            "properties": {
                "admins_identifiers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AccountIdentifiers"
                    }
                },
                "event": {
                    "type": "object",
                    "additionalProperties": true
                },
                "group_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "rest.createCalendarEventSingleGroupData": {
            "type": "object",
            "properties": {
                "event": {
                    "type": "object",
                    "additionalProperties": true
                },
                "to_members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ToMember"
                    }
                }
            }
        },
...
        "rest.updateCalendarEventSingleGroupData": {
            "type": "object",
            "properties": {
                "event": {
                    "type": "object",
                    "additionalProperties": true
                },
                "to_members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ToMember"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "type": "apiKey",
            "name": "ROKWIRE-API-KEY",
            "in": "header"
        },
        "AppUserAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header (add Bearer prefix to the Authorization value)"
        },
        "IntAPIKeyAuth": {
            "type": "apiKey",
            "name": "INTERNAL-API-KEY",
            "in": "header"
        }
    }
}